rimossa registrazione al df di ogni agente; comunicazione dello status funzionante;

+++ SLAsh/src/slash/entity/Status.java
@@ -1,40 +1,40 @@
 package slash.entity;
 
 import java.io.Serializable;
-import java.util.*;
-import slash.entity.*;
+import java.util.LinkedList;
+import java.util.List;
 
 public class Status implements Serializable {
 
 	private static final long serialVersionUID = -4467088461237807527L;
 
-	private List<Integer> latencyList;
+	private List<Float> latencyList;
 	private List<Float> reliabilityList;
-	private List<Integer> reqIntervalList;
+	private List<Float> reqIntervalList;
 	
-	private int latency;
+	private float latency;
 	private float reliability;
-	private int reqInterval;
+	private float reqInterval;
 	
 	public Status() {
-		this.latencyList = new LinkedList<Integer>();
+		this.latencyList = new LinkedList<Float>();
 		this.reliabilityList = new LinkedList<Float>();
-		this.reqIntervalList = new LinkedList<Integer>();
+		this.reqIntervalList = new LinkedList<Float>();
 	}
 	
-	public void addLatencyValue(int latencyValue) {
-		this.latencyList.add(Integer.valueOf(latencyValue));
+	public void addLatencyValue(float latencyValue) {
+		this.latencyList.add(Float.valueOf(latencyValue));
 	}
 	
 	public void addReliabilityValue(float reliabilityValue) {
 		this.reliabilityList.add(Float.valueOf(reliabilityValue));
 	}
 	
-	public void addReqIntervalValue(int reqIntervalValue) {
-		this.reqIntervalList.add(Integer.valueOf(reqIntervalValue));
+	public void addReqIntervalValue(float reqIntervalValue) {
+		this.reqIntervalList.add(Float.valueOf(reqIntervalValue));
 	}
 	
-	public int getAvgLatency() {
+	public float getAvgLatency() {
 		latency = 0;
 		for(int i=0; i<this.latencyList.size(); i++) {
 			latency += this.latencyList.get(i);
@@ -52,7 +52,7 @@ public class Status implements Serializable {
 		return reliability;
 	}
 	
-	public int getAvgReqInterval() {
+	public float getAvgReqInterval() {
 		reqInterval = 0;
 		for(int i=0; i<this.reqIntervalList.size(); i++) {
 			reqInterval += this.reqIntervalList.get(i);
